Visual-centric-AI-platform-for-effective-Learning-main/backend/requirements.txt
azure-ai-inference
azure-cognitiveservices-speech
azure-core
fastapi
google-api-python-client
google-generativeai
langchain
langchain-community
langchain-google-genai
PyPDF2
pydantic
python-docx
python-dotenv
pyttsx3
SpeechRecognition
supabase
uvicorn
python-multipart
langchain-core
httpx